+++ common/context/userContext.go
@@ -2,4 +2,5 @@ package context
 
 type UserContext struct {
 	Id uint64
+	Name string
 }
