+++ douyinService/resource/config.yaml
@@ -1,7 +1,7 @@
 redis:
   db: 0
-  addr: 47.96.157.235:6379
-  password: '000415'
+  addr: 127.0.0.1:6379
+  password: ''
 
 mysql:
   path: localhost
