+++ douyinService/router/router.go
@@ -2,7 +2,7 @@ package router
 
 import (
 	"github.com/DouYin/service/controller"
-	"github.com/DouYin/service/utils"
+	"github.com/DouYin/service/middleware"
 	"github.com/gin-gonic/gin"
 )
 
@@ -16,8 +16,8 @@ func (rt *Router) InitRouter(r *gin.RouterGroup) {
 	apiRouter := r.Group("/douyin")
 
 	//// basic apis
-	apiRouter.GET("/feed/", utils.JwtMiddleware(), controller.Feed)
-	apiRouter.GET("/user/", utils.JwtMiddleware(), controller.UserInfo)
+	apiRouter.GET("/feed/", middleware.JwtMiddleware(), controller.Feed)
+	apiRouter.GET("/user/", middleware.JwtMiddleware(), controller.UserInfo)
 	apiRouter.POST("/user/register/", controller.Register)
 	apiRouter.POST("/user/login/", controller.Login)
 	apiRouter.POST("/publish/action/", controller.Publish)
